# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MicroLens Recommendation MVP API",
    "description": "Phase 2D public runtime. Feed orchestration remains explicitly deferred to Phase 4.",
    "version": "0.2.0-phase2d",
    "x-implementation-status": "phase_2d_public_runtime"
  },
  "paths": {
...
    "/api/admin/dashboard/overview": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Overview",
        "operationId": "getDashboardOverview",
        "parameters": [
          {
            "name": "from_utc",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "From Utc"
            }
          },
          {
            "name": "to_utc",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "To Utc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardOverview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-required-roles": [
          "operator_readonly",
          "operator",
          "admin"
        ]
      }
    },
    "/api/admin/dashboard/timeseries": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Timeseries",
        "operationId": "getDashboardTimeseries",
        "parameters": [
          {
            "name": "from_utc",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "From Utc"
            }
          },
          {
            "name": "to_utc",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "To Utc"
            }
          },
          {
            "name": "feed_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FeedType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feed Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DashboardBucket"
                  },
                  "title": "Response Getdashboardtimeseries"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-required-roles": [
          "operator_readonly",
          "operator",
          "admin"
        ]
      }
    },
    "/api/admin/dashboard/feeds": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Feeds",
        "operationId": "getDashboardFeeds",
        "parameters": [
          {
            "name": "from_utc",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "From Utc"
            }
          },
          {
            "name": "to_utc",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "To Utc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardFeedDiagnostics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-required-roles": [
          "operator_readonly",
          "operator",
          "admin"
        ]
      }
    },
    "/api/admin/dashboard/export.csv": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Export Csv",
        "operationId": "exportDashboardCsv",
        "parameters": [
          {
            "name": "from_utc",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "From Utc"
            }
          },
          {
            "name": "to_utc",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "To Utc"
            }
          },
          {
            "name": "feed_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FeedType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feed Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-required-roles": [
          "operator_readonly",
          "operator",
          "admin"
        ]
      }
    },
    "/api/admin/dashboard/hot-items": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Hot Items",
        "operationId": "getDashboardHotItems",
        "parameters": [
          {
            "name": "from_utc",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "From Utc"
            }
          },
          {
            "name": "to_utc",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "To Utc"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HotItem"
                  },
                  "title": "Response Getdashboardhotitems"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-required-roles": [
          "operator_readonly",
          "operator",
          "admin"
        ]
      }
    },
    "/api/admin/users/{user_id}/debug": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Debug User",
        "operationId": "debugUser",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDebugResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-required-roles": [
          "operator_readonly",
          "operator",
          "admin"
        ]
      }
    },
    "/api/admin/requests/{request_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Debug Request",
        "operationId": "debugRecommendationRequest",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationRequestDebugResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-required-roles": [
          "operator_readonly",
          "operator",
          "admin"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DashboardBucket": {
        "properties": {
          "bucket_start_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Bucket Start Utc"
          },
          "bucket_end_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Bucket End Utc"
          },
          "feed_type": {
            "type": "string",
            "enum": [
              "personalized",
              "popular",
              "explore"
            ],
            "title": "Feed Type"
          },
          "request_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Request Count"
          },
          "exposure_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Exposure Count"
          },
          "click_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Click Count"
          },
          "like_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Like Count"
          },
          "share_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Share Count"
          },
          "revisit_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Revisit Count"
          },
          "dwell_ms_total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Dwell Ms Total"
          },
          "dwell_ms_avg": {
            "type": "number",
            "minimum": 0.0,
            "title": "Dwell Ms Avg"
          },
          "ctr": {
            "type": "number",
            "minimum": 0.0,
            "title": "Ctr"
          },
          "active_user_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Active User Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "bucket_start_utc",
          "bucket_end_utc",
          "feed_type",
          "request_count",
          "exposure_count",
          "click_count",
          "like_count",
          "share_count",
          "revisit_count",
          "dwell_ms_total",
          "dwell_ms_avg",
          "ctr",
          "active_user_count"
        ],
        "title": "DashboardBucket"
      },
      "DashboardFeedDiagnostics": {
        "properties": {
          "from_utc": {
            "type": "string",
            "format": "date-time",
            "title": "From Utc"
          },
          "to_utc": {
            "type": "string",
            "format": "date-time",
            "title": "To Utc"
          },
          "feeds": {
            "items": {
              "$ref": "#/components/schemas/DashboardBucket"
            },
            "type": "array",
            "title": "Feeds"
          },
          "feed_share": {
            "additionalProperties": {
              "type": "number"
            },
            "propertyNames": {
              "enum": [
                "personalized",
                "popular",
                "explore"
              ]
            },
            "type": "object",
            "title": "Feed Share"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "from_utc",
          "to_utc",
          "feeds",
          "feed_share"
        ],
        "title": "DashboardFeedDiagnostics"
      },
      "DashboardOverview": {
        "properties": {
          "from_utc": {
            "type": "string",
            "format": "date-time",
            "title": "From Utc"
          },
          "to_utc": {
            "type": "string",
            "format": "date-time",
            "title": "To Utc"
          },
          "timezone": {
            "type": "string",
            "const": "UTC",
            "title": "Timezone",
            "default": "UTC"
          },
          "total_users": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Users"
          },
          "active_users": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Active Users"
          },
          "requests": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Requests"
          },
          "exposures": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Exposures"
          },
          "clicks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Clicks"
          },
          "likes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Likes"
          },
          "shares": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Shares"
          },
          "revisits": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Revisits"
          },
          "dwell_ms_total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Dwell Ms Total"
          },
          "offline_item_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offline Item Count"
          },
          "ctr": {
            "type": "number",
            "minimum": 0.0,
            "title": "Ctr"
          },
          "zero_denominator": {
            "type": "boolean",
            "title": "Zero Denominator"
          },
          "active_model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Model Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "from_utc",
          "to_utc",
          "total_users",
          "active_users",
          "requests",
          "exposures",
          "clicks",
          "likes",
          "shares",
          "revisits",
          "dwell_ms_total",
          "offline_item_count",
          "ctr",
          "zero_denominator",
          "active_model_version"
        ],
        "title": "DashboardOverview"
      },
      "ErrorEnvelope": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id",
          "details"
        ],
        "title": "ErrorEnvelope"
      },
...
      "FeedItemResponse": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "cover": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover"
          },
          "position": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Position"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "item_id",
          "title",
          "cover",
          "position",
          "source",
          "score",
          "reason",
          "model_version"
        ],
        "title": "FeedItemResponse"
      },
...
      "FeedType": {
        "type": "string",
        "enum": [
          "personalized",
          "popular",
          "explore"
        ],
        "title": "FeedType"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HotItem": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "exposure_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Exposure Count"
          },
          "click_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Click Count"
          },
          "like_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Like Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "item_id",
          "title",
          "exposure_count",
          "click_count",
          "like_count"
        ],
        "title": "HotItem"
      },
...
      "RecommendationRequestDebugResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "candidate_item_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Candidate Item Ids"
          },
          "filtered_item_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Filtered Item Ids"
          },
          "ranked_items": {
            "items": {
              "$ref": "#/components/schemas/FeedItemResponse"
            },
            "type": "array",
            "title": "Ranked Items"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/PersistedEvent"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "candidate_item_ids",
          "filtered_item_ids",
          "ranked_items",
          "events"
        ],
        "title": "RecommendationRequestDebugResponse"
      },
...
      "UserDebugResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "profile": {
            "$ref": "#/components/schemas/UserProfileResponse"
          },
          "recent_request_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Recent Request Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_id",
          "profile",
          "recent_request_ids"
        ],
        "title": "UserDebugResponse"
      },
      "UserProfileResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "profile_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Profile Version"
          },
          "recent_interactions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Recent Interactions"
          },
          "positive_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Positive Summary"
          },
          "negative_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Negative Summary"
          },
          "dwell_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Dwell Summary"
          },
          "revisit_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Revisit Summary"
          },
          "share_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Share Summary"
          },
          "title_preferences": {
            "additionalProperties": true,
            "type": "object",
            "title": "Title Preferences"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_id",
          "profile_version",
          "recent_interactions",
          "positive_summary",
          "negative_summary",
          "dwell_summary",
          "revisit_summary",
          "share_summary",
          "title_preferences",
          "updated_at"
        ],
        "title": "UserProfileResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "microlens_session"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    },
    "responses": {
      "Error": {
        "description": "Structured API error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Host-published public listener"
    }
  ],
...
}